TRIPS-8 - Refactor backend code base
* Move api controllers into controllers module
