+++ main.go
@@ -7,7 +7,7 @@ import (
     "gopkg.in/mgo.v2"
     goWebServer "github.com/oskarszura/gowebscaffolding/gowebserver"
     "github.com/oskarszura/gowebscaffolding/controllers"
-    "github.com/oskarszura/gowebscaffolding/apicontrollers"
+    "github.com/oskarszura/gowebscaffolding/controllers/api"
     "github.com/oskarszura/gowebscaffolding/utils"
 )
 
@@ -29,7 +29,7 @@ func main() {
     log.Println("Connecting to mgo with mongoUri = " + mongoUri)
 
     if err != nil {
-        log.Fatal(err)
+        panic(err)
     }
 
     session, err := mgo.Dial(mongoUri)
@@ -37,27 +37,19 @@ func main() {
         panic(err)
     }
     defer session.Close()
+
     session.SetMode(mgo.Monotonic, true)
 
     utils.SetSession(session)
 
-    // Login
     server.Router.AddRoute("/login/register", controllers.Register)
     server.Router.AddRoute("/login/logout", controllers.AuthenticateLogout)
     server.Router.AddRoute("/login", controllers.Authenticate)
-
-    // Trips
     server.Router.AddRoute("/trips", controllers.Trips)
-
-    // Front
     server.Router.AddRoute("/", controllers.Front)
-
-    // Api
-    server.Router.AddRoute("/api/trips/{id}", apicontrollers.Trips)
-    server.Router.AddRoute("/api/places", apicontrollers.Places)
+    server.Router.AddRoute("/api/trips/{id}", api.Trips)
+    server.Router.AddRoute("/api/places", api.Places)
     server.Router.AddRoute("/api", controllers.Api)
-
-    // Errors
     server.Router.AddPageNotFoundRoute(controllers.NotFound)
 
     server.RunServer(addr)
