+++ models/trip.go
